+++ Assets/UtilityScripts/com.dman.reactive-variables/Runtime/BooleanVariable.cs
@@ -2,7 +2,7 @@
 
 namespace Dman.ReactiveVariables
 {
-    [CreateAssetMenu(fileName = "BooleanVariable", menuName = "ReactiveVariables/BooleanVariable", order = 2)]
+    [CreateAssetMenu(fileName = "BooleanVariable", menuName = "DmanUtilities/ReactiveVariables/BooleanVariable", order = 2)]
     public class BooleanVariable : GenericVariable<bool>
     {
     }