+++ Assets/UtilityScripts/com.dman.reactive-variables/Runtime/FloatVariable.cs
@@ -2,7 +2,7 @@
 
 namespace Dman.ReactiveVariables
 {
-    [CreateAssetMenu(fileName = "FloatVariable", menuName = "ReactiveVariables/FloatVariable", order = 2)]
+    [CreateAssetMenu(fileName = "FloatVariable", menuName = "DmanUtilities/ReactiveVariables/FloatVariable", order = 2)]
     public class FloatVariable : GenericVariable<float>
     {
         public float Add(float extraValue)