+++ Assets/UtilityScripts/com.dman.reactive-variables/Runtime/GameObjectVariable.cs
@@ -2,7 +2,7 @@
 
 namespace Dman.ReactiveVariables
 {
-    [CreateAssetMenu(fileName = "GameObjectVariable", menuName = "ReactiveVariables/GameObjectVariable", order = 1)]
+    [CreateAssetMenu(fileName = "GameObjectVariable", menuName = "DmanUtilities/ReactiveVariables/GameObjectVariable", order = 1)]
     public class GameObjectVariable : GenericVariable<GameObject>
     {
     }