+++ Assets/UtilityScripts/com.dman.reactive-variables/Runtime/IntVariable.cs
@@ -2,7 +2,7 @@
 
 namespace Dman.ReactiveVariables
 {
-    [CreateAssetMenu(fileName = "IntVariable", menuName = "ReactiveVariables/IntVariable", order = 2)]
+    [CreateAssetMenu(fileName = "IntVariable", menuName = "DmanUtilities/ReactiveVariables/IntVariable", order = 2)]
     public class IntVariable : GenericVariable<int>
     {
         public float Add(int extraValue)