+++ Assets/UtilityScripts/com.dman.scene-save-system/Runtime/Components/SaveablePrefabRegistry.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 namespace Dman.SceneSaveSystem
 {
-    [CreateAssetMenu(fileName = "SaveablePrefabRegistry", menuName = "Saving/SaveablePrefabRegistry", order = 10)]
+    [CreateAssetMenu(fileName = "SaveablePrefabRegistry", menuName = "DmanUtilities/Saving/SaveablePrefabRegistry", order = 10)]
     public class SaveablePrefabRegistry : UniqueObjectRegistryWithAccess<SaveablePrefabType>
     {
     }