+++ Assets/UtilityScripts/com.dman.scene-save-system/Runtime/Components/SaveablePrefabType.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 namespace Dman.SceneSaveSystem
 {
-    [CreateAssetMenu(fileName = "SaveablePrefabType", menuName = "Saving/SaveablePrefabType", order = 1)]
+    [CreateAssetMenu(fileName = "SaveablePrefabType", menuName = "DmanUtilities/Saving/SaveablePrefabType", order = 1)]
     public class SaveablePrefabType : IDableObject
     {
         public SaveablePrefab prefab;