+++ Assets/UtilityScripts/com.dman.reactive-variables/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.1.1] - 2021-02-14
+
+Scope asset menus under DmanUtilities
+
+### Added
+
+- Scoped asset menus
+
 ## [0.1.0] - 2021-01-09
 
 Initial release.