+++ Assets/UtilityScripts/com.dman.scene-save-system/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.1.1] - 2021-02-24
+
+scope all asset menus under DmanUtilities
+
+### Added
+
+- Scoped asset menus
+
 ## [0.1.0] - 2021-01-10
 
 Initial release.