+++ Assets/UtilityScripts/com.dman.utilities/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.3.1] - 2021-02-24
+
+### Added
+
+- create asset menu for raycast group
+
 ## [0.3.0] - 2021-01-16
 
 mouse over raycast helper. remove isMouseDown check on the RaycastToObject method.